Revert promise changes
